Updated Noisy folder - via sync tool update

+++ jupyter/noisy_baseline/ResNet18_baseline.py
@@ -9,41 +9,16 @@ import matplotlib.pyplot as plt
 from utilities.rectified_adam import RAdam
 import numpy as np
 import sys
-from utilities.pythonDB import writeToDB, deleteExistingPrimaryKeyDB
+from utilities.pythonDB import writeToDB, deleteExistingPrimaryKeyDB, recordsExists
+from utilities.resnet import ResNet18, ResNet34
 from utilities.data_preprocessors import get_random_data, MetricsAfterEachEpoch
 
 global MAX_EPOCHS, MAX_BATCH_SIZE, architecture, label, keywords
 MAX_EPOCHS, MAX_BATCH_SIZE = 15, 1024
-architecture = 'LeNet5'
+architecture = 'ResNet18'
 label, keywords = 'noisy_bl-experiments', 'benchmark_testing'
     
-def model_extractor(activation_func):
-    # Creating a LeNet5 Classifier
-    model = Sequential()
-
-    #Instantiating Layer 1
-    model.add(Conv2D(6, kernel_size=(5, 5), strides=(1, 1), activation=activation_func, padding='valid'))
-    model.add(AveragePooling2D(pool_size=(2, 2), strides=(2, 2), name='pool1'))
-
-    # #Instantiating Layer 2
-    model.add(Conv2D(16, kernel_size=(5, 5), strides=(1, 1), activation=activation_func, padding='valid'))
-    model.add(MaxPool2D(pool_size=(2, 2), strides=(2, 2), name='pool2'))
-
-    # #Instantiating Layer 3
-    model.add(Conv2D(120, kernel_size=(5, 5), strides=(1, 1), activation=activation_func, padding='valid'))
-
-    model.add(Flatten())
-
-    #Instantiating Layer 4
-    model.add(Dense(84, activation=activation_func)) 
-
-    #Output Layer
-    model.add(Dense(10, activation='softmax'))
-    return model
-
-
-def diff_optimizer(curr_optimizer, activation_func, label=None, curr_epochs=MAX_EPOCHS):
-    model = model_extractor(activation_func)
+def diff_optimizer(curr_optimizer, model, activation_func, label=None, curr_epochs=MAX_EPOCHS):
     model.compile(loss="categorical_crossentropy", optimizer=curr_optimizer, validation_data=(x_val, y_val),
                   metrics=["accuracy", "mae", "mse"])
     history = model.fit(x_train, y_train, epochs=curr_epochs, batch_size=1024).history
@@ -56,11 +31,10 @@ def diff_optimizer(curr_optimizer, activation_func, label=None, curr_epochs=MAX_
         model.save("../../saved_models_noisy/{}_{}_{}.hdf5".format(curr_optimizer, activation_func, architecture))
     return score, history
 
-
 if __name__ == '__main__':
     MAX_TRAINING = 20000
     MAX_VALIDATION = 2000
-    MAX_TESTING = 4000
+    MAX_TESTING = 5000
     
     base = 'noisy_dataset/noisy/cifar_10_{}.npy'
     x_train_max, y_train = np.load(base.format('x_train')), np.load(base.format('y_train'))
@@ -78,21 +52,29 @@ if __name__ == '__main__':
     
     testing_optimizers = ['RAdam', 'Adam', 'NAdam', 'SGD']
     testing_activations = ['tanh', 'relu', 'selu']
+    
+    img_rows, img_cols, img_channels = x_train[0].shape
+    nb_classes = y_test[0].shape[0]
 
-    for optimizer in testing_optimizers:
-        for activation in testing_activations:
+    for activation in testing_activations:
+        model = ResNet18((img_rows, img_cols, img_channels), nb_classes, activation)
+        for optimizer in testing_optimizers:
+
+            if recordsExists((architecture, label, optimizer, activation)):
+                print ("Continuing for Optimizer = {} and Activation = {}".format(optimizer, activation))
+                continue
+
             loss, accuracy, mae, mse = {}, {}, {}, {}
             train_loss_dict, train_accuracy_dict, train_mae_dict, train_mse_dict = {}, {}, {}, {}
-            
-            if optimizer == 'RAdam':
-                score_returned, history = diff_optimizer(RAdam(lr=0.005), activation, label='RAdam')
-            else:
-                score_returned, history = diff_optimizer(optimizer, activation)
 
-            
+            if optimizer == 'RAdam':
+                score_returned, history = diff_optimizer(RAdam(lr=0.005), model, activation, label='RAdam')
+            else:
+                score_returned, history = diff_optimizer(optimizer, model, activation)
+
             bg = (score_returned[0], score_returned[1], score_returned[2], score_returned[3], history.get('loss'),\
               history.get('accuracy'), history.get('mae'), history.get('mae'))
-        
+
             #Delete Existing Primary Keys and then write to DB
             deleteExistingPrimaryKeyDB(optimizer, activation, architecture, label, keywords)
             writeToDB(optimizer, activation, architecture, bg, label, keywords)
